Implement a better presence protocol handler PresenceProtocol.
This handler is not client-only, like PresenceClientProtocol, and accepts
`sender` arguments to all methods for outgoing presence. The `*Received`
methods now expect a single argument that is a parsed representation of the
presence stanza.

Also included is a simple XMPP server example that connects via the
server-to-server protocol, accepting presence requests and echoing incoming
messages for all potential addresses at a domain.


git-svn-id: https://svn.ik.nu/wokkel/trunk@186 b33ecbfc-034c-dc11-8662-000475d9059e